disabling memory limit!
i think there's a bug with it. i was playing some time and in a moment i wasn't able to join on server till i restart nginx. 
and i think that's because of limit_rate, i think that limit 10MB/s is for all connect players not "per one player" 
so 10MB/s will reached pretty fast and i'm disabling it for now!